+++ pyzoo/zoo/zouwu/use-case/network_traffic/network_traffic_autots_multivariant_example.py
@@ -76,7 +76,8 @@ trainer = AutoTSTrainer(dt_col="datetime",
                         extra_features_col=None)
 
 look_back = (36, 84)
-from zoo.automl.common.util import train_val_test_split
+from zoo.zouwu.preprocessing.utils import train_val_test_split
+
 train_df, val_df, test_df = train_val_test_split(df,
                                                  val_ratio=0.1,
                                                  test_ratio=0.1,
